Add input control widget and dialog
The fader selection is now done with the dialog, and the timesequence uses the widget.
Effect input selection can also be done with the dialog.
Fixes #117.

+++ system/writer.cpp
@@ -189,7 +189,7 @@ void Writer::writeFixtureFunction(const FixtureFunction& fixtureFunction)
 {
 	startElement("fixture-function");
 	writeAttribute("name", fixtureFunction.Name());
-	writeAttribute("type", fixtureFunction.Type());
+	writeAttribute("type", (int) fixtureFunction.Type());
 	writeDmxChannel(fixtureFunction.FirstChannel());
 	endElement();
 }
